Fix bug in the "np" function.
Slicing the variable "s" when returning causes the link not to appear
properly since such slicing remove the closing square bracket. This
issue would not raise if there is a blank line after the link. So to
avoid adding a blank line after a link specifically at the end of a file
 such slicing on the variable "s" is removed.